refactor send msg and banned bot case

+++ src/main/java/com/khmyl/telegram/currency/bot/service/subs/impl/SubscriberServiceImpl.java
@@ -2,12 +2,14 @@ package com.khmyl.telegram.currency.bot.service.subs.impl;
 
 import com.khmyl.telegram.currency.bot.converter.SubscriberConverter;
 import com.khmyl.telegram.currency.bot.model.dto.SubscriberDto;
+import com.khmyl.telegram.currency.bot.quartz.scheduler.SubscriberScheduler;
 import com.khmyl.telegram.currency.bot.repository.SubscriberRepository;
 import com.khmyl.telegram.currency.bot.service.subs.SubscriberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Component
@@ -16,9 +18,12 @@ public class SubscriberServiceImpl implements SubscriberService {
    @Autowired
    private SubscriberRepository subscriberRepository;
 
+   @Autowired
+   private SubscriberScheduler subscriberScheduler;
+
    @Override
    public void add(SubscriberDto dto) {
-      if (subscriberRepository.existsById(dto.getId())) {
+      if (!subscriberRepository.existsById(dto.getId())) {
          subscriberRepository.save(SubscriberConverter.convert(dto));
       }
    }
@@ -35,7 +40,12 @@ public class SubscriberServiceImpl implements SubscriberService {
 
    @Override
    public void remove(Long id) {
-      subscriberRepository.deleteById(id);
+      if (Objects.nonNull(id)) {
+         getById(id).ifPresent(sub -> {
+            subscriberScheduler.deleteSubJobs(sub);
+            subscriberRepository.deleteById(id);
+         });
+      }
    }
 
 }
